Allow ksh() to run as kthread
The kernel shell has not been well used since implementing proper
preemptive multitasking, since it relies on the older model of polling
the UART for input. However, some functionality exposed by the kernel
shell is still desirable, especially for testing new drivers. Allow the
kernel shell to run in SYS mode as a kernel thread. This allows the
shell to be put to sleep while using its own version of the "getchar"
system call. The kernel shell may be accessed by exiting the user shell,
which leaves no more processes to schedule.