Initial implementation of virtio-blk-mmio driver
The next major project in SOS is to implement a file system. To do this,
we need at least one block device! So, implement the beginnings of a
driver for virtio-blk devices. This implementation includes a struct for
the memory mappings, as well as code which can initialize a virtio-blk
driver, and send a first read command. We can tell that the read command
is successful because an unhandled interrupt is triggered. This is
enough progress for this commit.