virtio-blk: implement write support
So far, virtio-blk device was capable of sending a read command, but not
displaying the result. Further, writes were not possible. Implement
write commands, and allow the virtio-blk interrupt handler to display
the result of read commands as they arrive.

As part of this, we also implement proper free descriptor tracking in
the virtqueue implementation, so that we can always access the next
available free descriptor in constant time.