Use a per-process kernel stack
Prior to this commit, the kernel stack was global. Context switching
relied heavily on dumping data into this stack. System calls had limited
capability to block (the sysret field was hacked in to allow this). As a
result, implementing a blocking system call like read or write was very
difficult.

This commit moves to a per-process kernel stack model. When processes
make system calls, their context is dumped to this kernel stack. When
system calls block, on behalf of a process, they fill out the context
field. Similarly, when a process is interrupted, its context is dumped
to `current->context`.

We stop short of enabling kernel preemption. Although technically, it
should be entirely possible to preempt system calls, this would open the
door to a lot of concurrency issues I may not be ready to deal with yet.