Cosmetic improvements to shell prompts
Previously it was not very clear what "exit" did in the user shell and
why you suddenly get a WARNING and a different prompt. Since I like
showing off and taking screenshots, ensure that the prompts inform the
reader that each shell is different and let them know when the shells
change.