Simplify and improve slab allocator
We had a page-oriented approach, which enabled the neat feature that we
didn't need to pass the struct slab * to the free functio. But that
functionality was never all that useful. On the other hand, this
page-oriented approach meant that *every* page managed by the allocator
had to have a header.

While at some point we'll need to be able to reclaim pages from a slab,
which will want this functionality, right now it's wasteful.