net: Do proper memory management for packets
So far, packets are statically allocated and not properly managed.
Replace this strategy with a slab allocator, which can be shared for
packets from RX or destined to TX. As part of this, we ensure that each
layer of the network stack frees packets as it handles/drops them.
Further, in the virtio-net driver, we include a packet pointer which
allows easier retrieval of a struct packet from a descriptor. As a
result, our txused handler can free packets, and the rxused handler can
allocate new packets to place back into the RX virtqueue.