Better support custom tools & document Ubuntu setup
Qemu 4.2 is not standard on most Ubuntu, and the GDB command will be
different on Ubuntu. Slightly reorganize the Makefile to allow all tools
to be specified as variables, which could be overridden by the command
line. Add a .gitignore'd include to the Makefile to allow developers to
store their settings permanently. Update documentation in README and add
documentation file for Ubuntu.