Add wait system and allow runproc to wait for completion
Previously, the runproc system call would return once the process had
been created. For convenience in the shell (at least), I wanted to have
a way to wait for the completion of the process. This is not completely
convenient... a full implementation of the "wait" system call family
would be the best (but very complex). So I opted for a flag to runproc()
which asks to block until the process completes.

To do this, create a kernel API in wait.h which allows processes to wait
for a particular event to happen.