udp: Fix broken ports with sockets
udp_send() should accept ports in network byte order, but it instead
tried to convert them to network byte order. This inexplicably worked
with the destination address of sockets, since I forgot to use htons()
in userspace, but was broken for the receive address of sockets. Remove
the htons() calls and update callers of udp_send().