format: Implement SET() as a static inline function
The SET() macro allows us to output bytes to a destination buffer
without worrying about whether it will overrun the buffer; null
termination and bounds checking are all handled. Previously this was
implemented as a macro, which forced the inclusion of code, which is
difficult to optimize.

I've wised up and implemented this all as a static inline function. This
may seem silly, but it saves a ton of code space. I'm not sure about
whether it's any more or less cycles though.

Before:
    $ size lib/format.o
       text    data     bss     dec     hex filename
       3272       0       0    3272     cc8 lib/format.o

After:
    $ size lib/format.o
       text    data     bss     dec     hex filename
       2408       0       0    2408     968 lib/format.o