Allow debugging integration test failures
In the same way that we can run "make debug" and "make gdb" in separate
terminals, this supports debugging a particular test:

    # first terminal
    make testdebug TEST="test_help"
    # second terminal
    make gdb

The test runner notices when it is run in debug mode, and on faults it
will pause to let you continue debugging. If a fault happens and you are
not debugging, we raise an exception rather than wait for timeout.