integrationtests: Add test to run demo
To do this, update the vm to return any unmatched text to a buffer so
that the subsequent "read_until()" won't miss any text, and we can count
the number of occurrences of processes exiting.

The test currently fails frequently due to a race condition. I am
working on resolving it.