startup: create pre_mmu() startup hook
Prior to this point, we had used -fPIC, which failed to create a truly
position independent kernel. Thus, executing C code before the MMU was
setup (and thus, before the code was mapped into the position it was
linked for) resulted in bad behavior.

However, I recently learned that what I really wanted was -fPIE. By
using this option, C code may execute in the pre_mmu environment. This
commit on its own does nothing, but it sets us up for an exciting system
where we configure the MMU in C rather than ASM.