Add board support
This commit formally introduces the concept of kernel configuration, and
boards. Kernel configuration is a simple header file, and board is a
configuration option which selects some initialization routines.

With this commit, the UART base address is a kernel configuration option
as well, and its initialization sequence is modified: pre-mmu, the uart
is initialized for early puts. After mmu, the uart is remapped into
virtual memory, and once the interrupt controller is initialized, we can
turn on IRQ.

This allows us to remove knowledge of the UART from kmem_init().