uart: statically declare spinsem
This is a weird one. builds.sr.ht was consistently failing. I debugged
it and determined that it got stuck in the spinlock guarding the UART
reader. I noticed the spinlock was never initialized to 1. It appears
that this happens due to the pre-mmu environment - one would assume that
the generated code to modify uart_sem was using an absolute address
(somehow) and so the wrong address got written. A bisect revealed that
the issue was introduced in 58c802e780a8 ("Add board support"), which is
the first commit to use uart_init() in the pre_mmu() environment.

The strange thing is that this only reproduces on Debian builds.sr.ht.
The main difference in that environment from mine is gcc 9.3.0 versus my
laptop suing gcc 10.1.0. qemu version is the same. This may have to do
with the generated code. Whether or note the code generated under 9.3.0
is a bug is less clear. In any case, we can simply statically initialize
this variable, which resolves the issue on the build system.

Fixes: 58c802e780a895beb30a853c0ff504170cd174ed