fat: implement write()
This is an ugly commit which lumps too many items together. But I'm not
submitting it for review anywhere, and so I'll accept the small shame in
return for less time splitting it out. The patch implements write() for
FAT files, and since this is the only file system, it implements write()
in the generic VFS layer as well. To do this:

1) Add in flags for open() which indicate read-only / write-only files.
2) Notice that issuing block commands for the different size of buffers
   of FAT (cluster, sector, etc) is a bit fragmented, so centralize the
   operation.
3) Add functionality for seeking to the end of the file (only for open()
   with O_APPEND).
4) Add functionality for updating the size of a file in a directory
   entry.
5) Finally, add the block write logic, which looks (shockingly) similar
   to block read logic!