fat: fixes and tests to fat_write()
This commit brings us to the point where fat_write() will execute
correctly, at least for the limited addline command case. Non-append
mode writes are known to not work. Repeated read/write is broken. But
we can at least write a file (with multiple clusters) and read it back
the same as it went in. This is progress.