mailbox: flush cache before writing message
I would have thought that a memory barrier is necessary (and maybe it
should be). But the real deciding factor here is flushing the cache line
for the message, before we write that message value to the mailbox
interface. This ensures that values written to the buffer get seen on
the memory bus before we write the buffer address to the mailbox.