kmem: Introduce v2 management of page tables
The original design of SOS memory management was not a good one. The
physical address space was managed by one memory allocator, and the
virtual address space by a different one, and then process virtual
address spaces by yet another one. Memory allocations were satisfied by
allocating from the physical address space, and then the virtual address
space, and then mapping the addresses together.

Here's a short list of reasons why this was bad:

- Unmapping memory is tricky to get right. It requires proper cache and
  TLB maintenance. The complexity was not visible on QEMU, but on real
  hardware I have encountered countless issues. While eliminating the
  virtual memory operations from general page management won't eliminate
  the maintenance requirements (or bugs), it will dramatically reduce
  their scope and make it easier to debug.
- The virtual address space was entirely unrelated to the physical
  address space, so converting between addresses was complex and
  required manually consulting the page tables. Converting physical
  addresses to virtual ones was impossible. kmem v2 allows simple vtop
  and ptov operations for normal kernel memory without consulting page
  tables.
- Kernel page tables required using 2-level page tables and mapping
  everything using small pages. Now, normal kernel pages are mapped
  using large (1MB) section descriptors in the first-level table. We can
  still use small pages for some kernel mappings and for userspace
  mappings.

The kmem v2 approach is:

1. A 2GiB / 2GiB user kernel split. This is the best split available
   which allows both (a) a higher half kernel, and (b) keeping the user
   and kernel page tables completely separate in TTBR0 and TTBR1
   respectively.
2. Kernel virtual memory is a direct mapping onto physical memory. At
   the top of kernel memory is a virtual memory range for dynamic
   mappings (similar to Linux's vmalloc range).

To do this, we need to include first level / second level MMU page
flags, and we need to do a variety of small changes to memory usage
across the codebase. This is a mega-commit for sure, and leaves a few
TODOs which will be cleaned up in follow-ups. As of this commit,
several (but not all) integration tests pass, so the kernel is bootable
on QEMU.