alloc: Fix bad left shift for addr
Seems like the .addr field of struct zone may not always be the full
width, so when you do a "hdr->zones[i] << PAGE_BITS", you lose the top
bits. This... makes sense in some way, since the addr field is a 20 or
52 bit integer. To fix this, make sure all shifting reads use a special
"zone_addr()" accessor macro, which automatically casts to the full size
integer before doing the left shift.