linker: Properly test for architecture
My hopeful use of preprocessor macros (#if MACRO=identifier) was quite
wrong. Instead, define a macro for each architecture (ARCH_NAME) and
test for that in the config header files.

Part of this is that... we need to #include the arch includes. But both
configvals.h needs to have only preprocessor directives (no C
statements) for the linker script. So separate out the arch preprocessor
directives into arch/config.h.

The whole result is a bit of a messy change but fixes an issue where the
linker addresses were getting set all wrong due to the ARCH variable.